% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOQ.R
\name{EOQ}
\alias{EOQ}
\title{EOQ model}
\usage{
EOQ(l, k, I, C)
}
\arguments{
\item{l}{Demand (per unit of time).}

\item{k}{Preparation cost (per order).}

\item{I}{Storage cost (per article).}

\item{C}{Cost of goods (per item).}
}
\value{
A list containing:
\itemize{
\item{Q: }{Optimal order quantity.}
\item{Z: }{Total cost.}
\item{T: }{Cycle length.}
\item{N: }{Number of orders.}
}
}
\description{
This functions provides the results of the well-known deterministic EOQ (Economic Order Quantity model)
}
\details{
This function implements the basic deterministic EOQ (Economic Order Quantity) model.
}
\section{Assumptions}{

\itemize{
\item{Order arrives instantly.}
\item{No stockout.}
\item{Constant rate of demand.}
}
}

\examples{
\donttest{
l <- 520  # Demand
k <- 10   # Preparation cost
I <- 0.2  # Storage cost per article
C <- 5    # Cost of goods per item

dat <- EOQ(l = l, k = k, I = I, C = C)
dat
plot(dat)

}

}
\author{
José Carlos Soage González \email{jsoage@uvigo.es}
}
