% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{EOQdis-package}
\alias{EOQdis}
\alias{EOQdis-package}
\title{Package ‘EOQdis’}
\description{
This package implements the deterministic EOQ (Economic Order Quantity) with functions for the model with and without discounts.
}
\details{
Documentation for package ‘EOQdis’ version 1.0

The Description file:
\itemize{
\item{Package: EOQdis}
\item{Version: 1.0}
\item{Maintainer: José Carlos Soage González \email{jsoage@uvigo.es}}
\item{License: GPL-2}
}

Index of help topics:
\itemize{
\item{EOQ: }{Provides the results of the well-known deterministic EOQ (Economic Order Quantity model)}
\item{EOQd: }{Provides an EOQ with discounts where the units purchased have the same reduction in price.}
\item{EOQpd: }{Provides an EOQ with discounts where the discount occurs for units purchased when a certain amount is reached.}
\item{plot.EOQ: }{S3 method to plot a EOQ object by using the generic plot function.}
}
}
\author{
José Carlos Soage González. \email{jsoage@uvigo.es}
}
